+++ src/common/help.rs
@@ -1,6 +1,9 @@
 use crate::color::spec_color;
+use crate::utils::str_from_utf8;
+use lazy_static::lazy_static;
 use std::io::{Result, Write};
-use termcolor::{Color, WriteColor};
+use std::sync::Mutex;
+use termcolor::{Buffer, Color, WriteColor};
 
 const DOUBLE_LINE_PREFIX: &str = "===";
 const SIMPLE_LINE_PREFIX: &str = "---";
@@ -14,7 +17,42 @@ const PRIMARY_COLOR: Color = Color::Yellow;
 const SECONDARY_COLOR: Color = Color::Cyan;
 const CODE_COLOR: Color = Color::Green;
 
-pub fn highlight_help<O: Write + WriteColor>(output: &mut O, text: &str) -> Result<()> {
+static mut STATIC_HIGHLIGHTS: Vec<String> = Vec::new();
+
+lazy_static! {
+    static ref STATIC_HIGHLIGHTS_MUTEX: Mutex<()> = Mutex::new(());
+    static ref STATIC_HIGHLIGHTS_ENABLED: bool = atty::is(atty::Stream::Stdout)
+        && std::env::args().any(|arg| arg == "-h" || arg == "--help");
+}
+
+pub fn highlight_static(text: &str) -> &str {
+    if *STATIC_HIGHLIGHTS_ENABLED {
+        match highlight_to_string(text) {
+            Ok(result) => unsafe {
+                // Well, this is ugly but the current usage should be actually safe:
+                // 1) It's used only by cli.rs to generate static strings for clap attributes.
+                // 2) Values are never modified after being pushed to vector.
+                let _lock = STATIC_HIGHLIGHTS_MUTEX.lock(); // Lock might not be actually needed in our case
+                STATIC_HIGHLIGHTS.push(result);
+                STATIC_HIGHLIGHTS
+                    .last()
+                    .expect("Expected at least one static highlight result")
+                    .as_str()
+            },
+            Err(_) => text,
+        }
+    } else {
+        text
+    }
+}
+
+fn highlight_to_string(text: &str) -> Result<String> {
+    let mut buffer = Buffer::ansi();
+    highlight(&mut buffer, text)?;
+    str_from_utf8(buffer.as_slice()).map(String::from)
+}
+
+pub fn highlight<O: Write + WriteColor>(output: &mut O, text: &str) -> Result<()> {
     let mut in_heading = false;
     let mut in_padded_block_after_line = false;
 
@@ -90,38 +128,45 @@ fn highlight_code<O: Write + WriteColor>(output: &mut O, line: &str) -> Result<(
 #[cfg(test)]
 mod tests {
     use crate::testing::{ColoredOuput, OutputChunk};
+    use claim::*;
     use indoc::indoc;
 
+    const SAMPLE_HELP: &str = indoc! {"
+        =========
+         Heading
+        =========
+
+        Text.
+
+        === Heading ===
+
+        Text.
+        Text with `code`.
+        Text with `code` and `more code`.
+
+            KEY    VALUE
+            ------------
+            `a`      1
+            `b`      2
+
+        Text.
+
+            $> ls -la
+            $> ls -la # Shell comment
+    "};
+
     #[test]
-    fn highlight_help() {
+    fn highlight_to_string() {
+        use super::*;
+        assert_gt!(highlight_to_string(SAMPLE_HELP).unwrap().len(), 0);
+    }
+
+    #[test]
+    fn highlight() {
         use super::*;
 
-        let help = indoc! {"
-            =========
-             Heading
-            =========
-
-            Text.
-
-            === Heading ===
-
-            Text.
-            Text with `code`.
-            Text with `code` and `more code`.
-            
-                KEY    VALUE
-                ------------
-                `a`      1
-                `b`      2
-
-            Text.
-
-                $> ls -la
-                $> ls -la # Shell comment
-        "};
-
         let mut ouput = ColoredOuput::new();
-        highlight_help(&mut ouput, help).unwrap();
+        highlight(&mut ouput, SAMPLE_HELP).unwrap();
         assert_eq!(
             ouput.chunks(),
             &[